# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pauza API
  description: Backend HTTP API for the Pauza focus-timer application.
  version: 0.1.0
  contact:
    name: Pauza Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.pauza.app
    description: Production (placeholder)
...
components:
  securitySchemes:
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token issued by POST /api/v1/auth/verify or POST /api/v1/auth/refresh.
    AdminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token issued by POST /api/v1/admin/login.
    WebhookBearer:
      type: http
      scheme: bearer
      description: Static secret shared with RevenueCat for webhook authentication.
  headers:
    XRequestId:
      description: Request identifier echoed in every response.
      schema:
        type: string
    XRateLimitLimit:
      description: Maximum requests allowed in the current window when the limiter is enforcing a real budget.
      schema:
        type: integer
    XRateLimitRemaining:
      description: Remaining requests in the current window when the limiter is enforcing a real budget.
      schema:
        type: integer
    XRateLimitReset:
      description: UTC Unix timestamp when the current window resets, when the limiter is enforcing a real budget.
      schema:
        type: integer

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    UserResponse:
      type: object
      required: [id, email, name, username, profile_picture_url, push_enabled, leaderboard_visible, created_at, subscription]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        username:
          type: string
        profile_picture_url:
          type: string
          nullable: true
        push_enabled:
          type: boolean
        leaderboard_visible:
          type: boolean
        created_at:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
        subscription:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/SubscriptionResponse"
...
    UpdateMeRequest:
      type: object
      additionalProperties: false
      description: Unknown JSON fields are rejected. All fields are optional.
      properties:
        name:
          type: string
          maxLength: 100
        username:
          type: string
          pattern: '^[A-Za-z0-9_]{3,30}$'
          description: 3-30 characters; letters, digits, and underscores only.
    PrivacyPreferencesResponse:
      type: object
      required: [leaderboard_visible]
      properties:
        leaderboard_visible:
          type: boolean
    UpdatePrivacyPreferencesRequest:
      type: object
      additionalProperties: false
      description: Unknown JSON fields are rejected. All fields are optional.
      properties:
        leaderboard_visible:
          type: boolean
    NotificationPreferencesResponse:
      type: object
      required: [push_enabled]
      properties:
        push_enabled:
          type: boolean
    UpdateNotificationPreferencesRequest:
      type: object
      additionalProperties: false
      description: Unknown JSON fields are rejected. All fields are optional.
      properties:
        push_enabled:
          type: boolean
    UsernameAvailableResponse:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
    PhotoUploadResponse:
      type: object
      required: [profile_picture_url]
      properties:
        profile_picture_url:
          type: string
          description: Public URL served by the deployment's reverse proxy from machine-local photo storage.
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    DeleteConfirmRequest:
      type: object
      additionalProperties: false
      required: [otp]
      description: Unknown JSON fields are rejected.
      properties:
        otp:
          type: string
          pattern: '^[0-9]{6}$'
          description: Exactly 6 ASCII digits.

    # -- Devices ------------------------------------------------------------
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request validation failed (422).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Resource conflict (e.g. duplicate username).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Too many requests. Rate-limit middleware also emits informational headers on enforced responses.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        X-RateLimit-Limit:
          $ref: "#/components/headers/XRateLimitLimit"
          required: false
        X-RateLimit-Remaining:
          $ref: "#/components/headers/XRateLimitRemaining"
          required: false
        X-RateLimit-Reset:
          $ref: "#/components/headers/XRateLimitReset"
          required: false
        Retry-After:
          schema:
            type: integer
          required: false
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Unexpected server error.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/me:
    get:
      operationId: getMe
      tags: [Profile]
      summary: Get current user profile
      description: Requires a user JWT. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      responses:
        "200":
          description: User profile returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateMe
      tags: [Profile]
      summary: Update current user profile
      description: Requires a user JWT. Unknown JSON fields are rejected. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Updated user profile returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/notification-preferences:
    get:
      operationId: getNotificationPreferences
      tags: [Profile]
      summary: Get current user notification preferences
      description: Requires a user JWT. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      responses:
        "200":
          description: Notification preferences returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updateNotificationPreferences
      tags: [Profile]
      summary: Update current user notification preferences
      description: Requires a user JWT. Unknown JSON fields are rejected. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesRequest"
      responses:
        "200":
          description: Updated notification preferences returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/privacy:
    get:
      operationId: getPrivacyPreferences
      tags: [Profile]
      summary: Get current user privacy preferences
      description: Requires a user JWT. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      responses:
        "200":
          description: Privacy preferences returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyPreferencesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updatePrivacyPreferences
      tags: [Profile]
      summary: Update current user privacy preferences
      description: Requires a user JWT. Unknown JSON fields are rejected. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrivacyPreferencesRequest"
      responses:
        "200":
          description: Updated privacy preferences returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyPreferencesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/username-available:
    get:
      operationId: usernameAvailable
      tags: [Profile]
      summary: Check if a username is available
      description: Requires a user JWT. Rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z0-9_]{3,30}$'
          description: 3-30 characters; letters, digits, and underscores only.
      responses:
        "200":
          description: Availability result.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAvailableResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/photo:
    post:
      operationId: uploadPhoto
      tags: [Profile]
      summary: Upload profile photo
      description: Accepts a JPEG or PNG image up to 5 MiB. This endpoint applies its own 5 MiB body cap (overriding the global 1 MiB default). Requires a user JWT and is rate limited by user ID using the `general-api` tier. Uploaded files are stored on machine-local disk and are expected to be served by a reverse proxy such as Nginx at the configured `PHOTO_PUBLIC_BASE_URL`.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        "200":
          description: Photo uploaded; new URL returned.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/delete/request:
    post:
      operationId: deleteRequest
      tags: [Profile]
      summary: Request account deletion
      description: Sends a confirmation OTP to the user's email. The message is intentionally vague for anti-enumeration. Requires a user JWT and is rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      responses:
        "200":
          description: 'Returns the generic anti-enumeration message, "If the account is eligible for deletion, a confirmation code has been sent."'
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/me/delete/confirm:
    post:
      operationId: deleteConfirm
      tags: [Profile]
      summary: Confirm account deletion
      description: Verifies the deletion OTP and permanently deletes the user account. Unknown JSON fields are rejected. Requires a user JWT and is rate limited by user ID using the `general-api` tier.
      security:
        - UserJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteConfirmRequest"
      responses:
        "200":
          description: Account deleted.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            X-RateLimit-Limit:
              $ref: "#/components/headers/XRateLimitLimit"
            X-RateLimit-Remaining:
              $ref: "#/components/headers/XRateLimitRemaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/XRateLimitReset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # Social — Devices
  # =========================================================================
...